4.5. Отладка программы в IDEA [#159464]

+++ src/test/java/ru/job4j/condition/SqAreaTest.java
@@ -9,7 +9,8 @@ public class SqAreaTest {
     public void whenP6K2Square2() {
         int expected = 2;
         int p = 6;
-        double k = 2;
+        //double k = 2;
+        int k = 2;
         double out = SqArea.square(p, k);
         Assert.assertEquals(expected, out, 0.01);
     }
@@ -18,7 +19,8 @@ public class SqAreaTest {
     public void whenP4K1Square1() {
         int expected = 1;
         int p = 4;
-        double k = 1;
+        //double k = 1;
+        int k = 1;
         double out = SqArea.square(p, k);
         Assert.assertEquals(expected, out, 0.01);
     }
@@ -27,7 +29,8 @@ public class SqAreaTest {
     public void whenP10K3Square46875() {
         double expected = 4.6875;
         int p = 10;
-        double k = 3;
+        //double k = 3;
+        int k = 3;
         double out = SqArea.square(p, k);
         Assert.assertEquals(expected, out, 0.0001);
     }
